+++ ants-cpp/Ant.h
@@ -11,11 +11,11 @@ class Ant
         vector<int> path;
         double alpha;
         double beta;
-        vector<vector<int>>* cities;
+        vector<vector<double>>* cities;
         vector<vector<vector<double>>>* cities_matrix;
         vector<int> unvisited_cities;
     public:
-        Ant(double alpha, double beta, vector<vector<int>>* cities, vector<vector<vector<double>>>* cities_matrix);
+        Ant(double alpha, double beta, vector<vector<double>>* cities, vector<vector<vector<double>>>* cities_matrix);
         int choose_city();
         void generate_path();
         void check_unvisited(); // return list of unvisited cities
