+++ ants-cpp/Colony.h
@@ -11,7 +11,7 @@ class Colony
         int ant_count;
         vector<Ant> ants;
         double rho;
-        vector<vector<int>>* cities;
+        vector<vector<double>>* cities;
         vector<vector<vector<double>>>* cities_matrix;
         int steps;
         double min_scalling_factor;
@@ -21,7 +21,7 @@ class Colony
     public:
         Colony(int ant_count, double min_s_f, double alpha,
             double beta, double rho, int steps,
-            vector<vector<int>>* cities, vector<vector<vector<double>>>* cities_matrix,
+            vector<vector<double>>* cities, vector<vector<vector<double>>>* cities_matrix,
             double pheromone_weight);
         void solve();
         void place_pheromones(vector<int> path, double distance);
